refactor: remove duplicate GroupPermissions type
