+++ src/auth/shield/shield.utils.ts
@@ -1,5 +1,5 @@
 import { UNAUTHORIZED } from "../../common/common.constants";
-import { GroupPermissions } from "../../groups/group-roles/models/group-role-permission.model";
+import { GroupPermissions } from "../../groups/group-roles/models/group-permissions.type";
 import { ServerPermissions } from "../../server-roles/models/server-permissions.type";
 import { UserPermissions } from "../../users/user.types";
