+++ src/dataloader/dataloader.types.ts
@@ -1,5 +1,5 @@
 import DataLoader from "dataloader";
-import { GroupPermissions } from "../groups/group-roles/models/group-role-permission.model";
+import { GroupPermissions } from "../groups/group-roles/models/group-permissions.type";
 import { Group } from "../groups/models/group.model";
 import { Image } from "../images/models/image.model";
 import { Like } from "../likes/models/like.model";
